skeleton loading for explore page
