+++ apps/client/app/loading.tsx
@@ -1,8 +1,8 @@
 'use client'
-import { OwnLoader } from '@/components/custom/loader'
+// import { OwnLoader } from '@/components/custom/loader'
 
 export default function Loading() {
   return (
-    <OwnLoader />
+    <p>Loading...</p>
   )
 }
