--- apps/client/app/testtingg/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Testing() {
-  return (
-    <div>
-      hey
-    </div>
-  )
-}
